Update benchmark thresholds -- a decent amount of malloc decrements
In 24a4f47 I unnested a nested function and simply used inout arguments.
I had bittern by the compiler boxing stuff before, for nested functions, when the box wasn't really needed, so I thought there is a chance this might help.
I'm a bit surprised, but it surely did go to the plan.